+++ posce/comms/base.py
@@ -12,8 +12,8 @@ SETTINGS = {
 }
 
 @click.group(cls=CustomGroup, context_settings=SETTINGS)
-@click.option('--dire', envvar='POSCE_DIR', hidden=True)
-@click.option('--ext',  envvar='POSCE_EXT', hidden=True)
+@click.option('--dir', envvar='POSCE_DIR', hidden=True)
+@click.option('--ext', envvar='POSCE_EXT', hidden=True)
 @click.pass_context
 def group(ctx, dire, ext):
     '''