+++ posce/comms/__init__.py
@@ -7,4 +7,5 @@ from posce.comms import copy
 from posce.comms import edit
 from posce.comms import find
 from posce.comms import list
+from posce.comms import move
 from posce.comms import show