+++ tests/test_comms/test_move.py
@@ -1,23 +1,23 @@
 '''
-Tests for 'posce.comms.copy'.
+Tests for 'posce.comms.move'.
 '''
 
 import os.path
 
-from posce.comms.copy           import copy
+from posce.comms.move           import move
 from tests.test_items.test_book import book
 from tests.tools                import out
 
-def test_copy(book):
+def test_move(book):
     # setup
     dest = book['alpha'].path.replace('alpha', 'dest')
 
     # success - defaults
-    assert out(book, copy, 'alpha', 'dest') == []
-    assert os.path.exists(dest)
-    assert os.path.exists(book['alpha'].path)
+    assert out(book, move, 'alpha', 'dest') == []
+    assert     os.path.exists(dest)
+    assert not os.path.exists(book['alpha'].path)
 
     # failure - dest exists
-    assert out(book, copy, 'bravo', 'charlie') == [
+    assert out(book, move, 'bravo', 'charlie') == [
         "Error: Note 'charlie' already exists.\n",
     ]