+++ posce/comms/__init__.py
@@ -4,6 +4,7 @@ Package definition for 'posce.comms'.
 
 from posce.comms import base
 from posce.comms import copy
+from posce.comms import dump
 from posce.comms import edit
 from posce.comms import find
 from posce.comms import list