+++ changes.md
@@ -2,8 +2,12 @@ Changelog
 =========
 
 ## Unreleased
+### Added
+- New commands: `copy`.
+
 ### Changed
 - Renamed `--dire` option to `--dir`.
+
 
 ## [0.2.0] · 2020-03-19
 ### Added