+++ changes.md
@@ -2,6 +2,9 @@ Changelog
 =========
 
 ## Unreleased
+
+
+## [0.3.0] · 2020-03-24
 ### Added
 - New commands: `copy`, `dump`, `move`, `wget`, `repl` (hidden).
 - New option `-v --version` to show current version.
@@ -36,6 +39,7 @@ Changelog
 
 
 [Unreleased]: https://github.com/posce/posce/commits/master
+[0.3.0]:      https://github.com/posce/posce/commits/0.3.0
 [0.2.0]:      https://github.com/posce/posce/commits/0.2.0
 [0.1.1]:      https://github.com/posce/posce/commits/0.1.1
 [0.1.0]:      https://github.com/posce/posce/commits/0.1.0